+++ spec/utils/config-helper.spec.js
@@ -12,9 +12,6 @@ describe("Config-helper ", function() {
     describe('parsePaths', function(){
 
         var paths = {
-            dist: {
-                root:'./dist'
-            },
             src: {
                 root:'./src'
             }
@@ -22,11 +19,6 @@ describe("Config-helper ", function() {
 
         it("parsePaths ensures the object always has scripts, styles, fonts, icons and images", function () {
             var newPaths = helper.parsePaths(paths)
-            expect(newPaths.dist.scripts).toBe('./dist/scripts');
-            expect(newPaths.dist.styles).toBe('./dist/styles');
-            expect(newPaths.dist.fonts).toBe('./dist/fonts');
-            expect(newPaths.dist.icons).toBe('./dist/icons');
-            expect(newPaths.dist.images).toBe('./dist/images');
             expect(newPaths.src.scripts).toBe('./src/scripts');
             expect(newPaths.src.styles).toBe('./src/styles');
             expect(newPaths.src.fonts).toBe('./src/fonts');
@@ -35,14 +27,8 @@ describe("Config-helper ", function() {
         });
 
         it("does not override given paths", function () {
-            paths.dist.scripts = './mypath/js'
             paths.src.images = './mypath/images-mofo'
             var newPaths = helper.parsePaths(paths)
-            expect(newPaths.dist.scripts).toBe('./mypath/js');
-            expect(newPaths.dist.styles).toBe('./dist/styles');
-            expect(newPaths.dist.fonts).toBe('./dist/fonts');
-            expect(newPaths.dist.icons).toBe('./dist/icons');
-            expect(newPaths.dist.images).toBe('./dist/images');
             expect(newPaths.src.scripts).toBe('./src/scripts');
             expect(newPaths.src.styles).toBe('./src/styles');
             expect(newPaths.src.fonts).toBe('./src/fonts');