+++ tasks/build.js
@@ -2,6 +2,7 @@ var Promise = require('es6-promise').Promise;
 var log = require('./utils/log');
 var fs = require('./utils/fs');
 var helper = require('./utils/config-helper');
+var clean = require('./clean');
 var config, paths, pkg;
 
 function initConfig(){
@@ -10,18 +11,18 @@ function initConfig(){
     pkg = config.pkg;
 }
 
-var clean = require('./clean');
 
+//todo: move to copy task
 function serverConfigFiles(){
     initConfig();
     var source = paths.source.root +'/' + '{CNAME,.htaccess,robots.txt}';
     return fs.copy(source, paths.site.root);
 }
+
 function html(replacements) {
     initConfig();
     var build = helper.matches(config.build, ['jade','mustache']);
     if (!build || !paths.site){
-        log.info('Skipping build html');
         return Promise.resolve();
     }
     replacements = replacements || {};
@@ -51,28 +52,28 @@ function html(replacements) {
         });
         return Promise.all(promises);
     }).then(function(){
-        log.info('Build HTML Complete');
+        log.info(' * HTML Complete');
     }).catch(log.warn);
 }
 
+//todo: move to copy task
 function fonts() {
     initConfig();
     var build = helper.matches(config.build, ['fonts']);
     if (!build || !paths.site) {
-        log.info('skipping build fonts');
         return Promise.resolve();
     }
     var location = [];
     paths.source && paths.source.fonts && location.push(paths.source.fonts + '/**/*.{eot,ttf,woff,svg}');
-    paths.bower && paths.bower.fonts && location.push(paths.bower.fonts + '/**/*.{eot,ttf,woff,svg}');
+    paths.demo && paths.demo.fonts && location.push(paths.demo.fonts + '/**/*.{eot,ttf,woff,svg}');
     return fs.copy(location, paths.site.fonts);
 }
 
+//todo: move to copy task
 function images() {
     initConfig();
     var build = helper.matches(config.build, ['images']);
     if (!build || !paths.site) {
-        log.info('skipping build images');
         return Promise.resolve();
     }
     var location = [];
@@ -85,7 +86,6 @@ function scripts(options){
     initConfig();
     var build = helper.matches(config.build, ['browserify','requirejs']);
     if (!build){
-        log.info('skipping build scripts');
         return Promise.resolve();
     }
     var Scripts = require('./wrappers/' + build);
@@ -94,11 +94,10 @@ function scripts(options){
     options.browser = pkg.browser;
     options["browserify-shim"] = pkg["browserify-shim"];
     return Promise.all([
-        paths.dist && paths.dist.scripts && new Scripts(paths.source.scripts, paths.dist.scripts, options).write(),
         paths.demo && paths.demo.scripts && new Scripts(paths.demo.scripts, paths.site.scripts, options).write(),
         paths.site && paths.site.scripts && new Scripts(paths.source.scripts, paths.site.scripts, options).write()
     ]).then(function(){
-        log.info('Build Scripts Complete');
+        log.info(' * Scripts Complete');
     }).catch(log.warn);
 }
 
@@ -106,23 +105,21 @@ function styles(options){
     initConfig();
     var build = helper.matches(config.build, ['sass']);
     if (!build){
-        log.info('Skipping build Sass');
         return Promise.resolve();
     }
     var Styles = require('./wrappers/' + build);
     options = options || (config[build]) || {};
     return Promise.all([
-        paths.dist && paths.dist.styles && new Styles(paths.source.styles, paths.dist.styles, options).write(),
         paths.site && paths.site.styles && new Styles(paths.source.styles, paths.site.styles, options).write(),
         paths.demo && paths.demo.styles && new Styles(paths.demo.styles, paths.site.styles, options).write()
     ]).then(function(){
-        log.info('Build Styles Complete');
+        log.info(' * Styles Complete');
     }).catch(log.warn);
 }
 
 function run(replacements){
-    return clean.all().then(function(){
-        log.info('Build :');
+    return clean('all').then(function(){
+        log.info('Building :');
         return Promise.all([
                 serverConfigFiles(),
                 scripts(),
@@ -131,8 +128,6 @@ function run(replacements){
                 styles(),
                 html(replacements)
             ]);
-    }).then(function(){
-        log.info('Build All Complete');
     }).catch(log.warn);
 }
 
@@ -143,6 +138,6 @@ module.exports = {
     scripts: scripts,
     images: images,
     fonts: fonts,
-    run: run,
-    all: run
+    run: run,//todo: choose run or all, not both!
+    all: run//todo: choose run or all, not both!
 };